requirements/base.txt
Django==1.8.3
psycopg2==2.6.1
pytz==2015.4
django-admin-bootstrapped==2.5.4
djangorestframework==3.1.3
PyJWT==1.4.0
django-pipeline==1.5.2
django-filter==0.10.0
raven==5.7.1
django-rest-swagger==0.3.4
cryptography==1.1
prometheus-client==0.0.13
django-prometheus==1.0.4
requests==2.9.1
